+++ mypyc/irbuild/function.py
